requirements.txt
websockets>=11.0
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
PyQt5>=5.15
scipy>=1.10
pyyaml>=6.0